Feature/backend user controller

See merge request s05-final/S05P31A402!24

+++ backend/src/main/java/com/ssacretary/db/entity/Log.java
@@ -18,7 +18,7 @@ public class Log {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "log_id")
-    private Long logId;
+    private int logId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "setting_id")
